requirements.txt
entsoe-py
requests
pandas
numpy
matplotlib
seaborn
scikit-learn
streamlit
plotly